Refactor code so that stage endpoint returns the public image url
Because the consumer of the "stage" endpoint is interested in the image
url as something they can use to pull the image. Up to now, what we
returned was only useful to Kubernetes, because it could access the
registry through 127.0.0.1.

We now replace the URL in the "deploy" endpoint as a preparation to use
and external registry as the "Epinio registry".

Co-authored-by: Dimitris Karakasilis <DKarakasilis@suse.com>