Move `use-internal-registry-node-port` to `force-kube-internal-registry-tls`
with a default value of "false". This flag will be ignored if an
external registry is used (TODO). A warning will be shown in that case.

We rename the flag to make more sense to the user. From the user's
perspective, the reason to set it to true is to make sure the traffic
between the registry and Kubernetes is encrypted. The old name was
simply revealing how we handle that case internally but made no sense to
the user.